- The ParameterNotInitializedException if no longer thrown for optional     parameter or attributes in the request. If you have an optional parameter     or attribute in the input, an object will always be returned from the get     method. The value of this object will be null if the parameter is not     sent and there is no default.     This change may break the compatibility of your code if you had optional     input parameters or attributes returning primitive types (int, short,     boolean, ...) you will need to change your code to use .intValue(), etc...

+++ src/tests/apis/allinone/impl/com/mycompany/allinone/api/AttributeComboImpl.java
@@ -71,7 +71,7 @@ public final class AttributeComboImpl extends AttributeCombo {
          Person nextPerson = (Person)itPersons.next();
          Registration registration = new Registration();
          if (nextPerson.isSetAge()) {
-            int age = nextPerson.getAge();
+            int age = nextPerson.getAge().intValue();
             registration.setRegistrationYear(calendar.get(Calendar.YEAR) - age + 1);
             registration.setRegistrationMonth(calendar.get(Calendar.MONTH));
          } else {
@@ -83,9 +83,9 @@ public final class AttributeComboImpl extends AttributeCombo {
                month = nextPerson.getBirthDate().getMonthOfYear();
                day = nextPerson.getBirthDate().getDayOfMonth();
             } else {
-               year = nextPerson.getBirthYear();
-               month = nextPerson.getBirthMonth();
-               day = nextPerson.getBirthDay();
+               year = nextPerson.getBirthYear().intValue();
+               month = nextPerson.getBirthMonth().intValue();
+               day = nextPerson.getBirthDay().intValue();
             }
 
             // Create an invalid response for year above 2006
